# docs/swagger.yaml
definitions:
  container.ApiResponse:
    properties:
      data: {}
      message:
        type: string
      status:
        description: success | fail
        type: string
    type: object
  container.CreateContainerRequest:
    properties:
      command:
        example:
        - '''/bin/sh'''
        - '''-c'''
        - '''echo hello; sleep 60'''
        items:
          type: string
        type: array
      image:
        example: alpine:latest
        type: string
      mount:
        example:
        - '''/host/dir:/container/dir'''
        - '''/src:/dst'''
        items:
          type: string
        type: array
      port:
        example:
        - '''8080:80'''
        - '''4443:443'''
        items:
          type: string
        type: array
    type: object
  container.ExecContainerRequest:
    properties:
      command:
        example:
        - '''/bin/sh'''
        - '''-c'''
        - '''echo hello'''
        items:
          type: string
        type: array
      interactive:
        example: true
        type: boolean
    type: object
  container.StartContainerRequest:
    properties:
      interactive:
        example: true
        type: boolean
    type: object
  http.ApiResponse:
    properties:
      data: {}
      message:
        type: string
      status:
        description: success | fail
        type: string
    type: object
  image.ApiResponse:
    properties:
      data: {}
      message:
        type: string
      status:
        description: success | fail
        type: string
    type: object
  image.PullImageRequest:
    properties:
      arch:
        example: arm64
        type: string
      image:
        example: alpine:latest
        type: string
      os:
        example: linux
        type: string
    type: object
  image.RemoveImageRequest:
    properties:
      image:
        example: alpine:latest
        type: string
    type: object
info:
  contact: {}
paths:
  /v1/containers:
    post:
      consumes:
      - application/json
      description: create a new container
      parameters:
      - description: Container Spec
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/container.CreateContainerRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/container.ApiResponse'
      summary: Create a container
      tags:
      - containers
  /v1/containers/{containerId}/actions/delete:
    delete:
      description: delete an exitsting container
      parameters:
      - description: Container ID
        in: path
        name: containerId
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/container.ApiResponse'
      summary: delete a container
      tags:
      - containers
  /v1/containers/{containerId}/actions/exec:
    post:
      description: execute command inside an exitsting container
      parameters:
      - description: Container ID
        in: path
        name: containerId
        required: true
        type: string
      - description: Execute Options
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/container.ExecContainerRequest'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/container.ApiResponse'
      summary: exec a container
      tags:
      - containers
  /v1/containers/{containerId}/actions/start:
    post:
      description: start an exitsting container
      parameters:
      - description: Container ID
        in: path
        name: containerId
        required: true
        type: string
      - description: Start Options
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/container.StartContainerRequest'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/container.ApiResponse'
      summary: start a container
      tags:
      - containers
  /v1/containers/{containerId}/actions/stop:
    post:
      description: stop an exitsting container
      parameters:
      - description: Container ID
        in: path
        name: containerId
        required: true
        type: string
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/container.ApiResponse'
      summary: stop a container
      tags:
      - containers
  /v1/hooks/droplet:
    post:
      description: apply hook from droplet
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.ApiResponse'
      summary: apply hook
      tags:
      - Hooks
  /v1/images:
    delete:
      consumes:
      - application/json
      description: remove image from local
      parameters:
      - description: Target Image
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/image.RemoveImageRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/image.ApiResponse'
      summary: remove image
      tags:
      - image
    post:
      consumes:
      - application/json
      description: pull image from registry
      parameters:
      - description: Target Image
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/image.PullImageRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/image.ApiResponse'
      summary: pull image
      tags:
      - image
swagger: "2.0"